requirements.txt
absl-py==1.4.0
addict==2.4.0
angles==1.9.12
anyio==3.6.2
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
attrs==22.2.0
backcall==0.2.0
beautifulsoup4==4.11.2
black==23.1.0
bleach==6.0.0
bondpy==1.8.5
brotlipy==0.7.0
cachetools==5.3.0
camera_calibration==1.15.2
camera_calibration_parsers==1.11.13
catkin==0.7.29
certifi @ file:///croot/certifi_1671487769961/work/certifi
cffi @ file:///croot/cffi_1670423208954/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click==8.1.3
colorama==0.4.6
comm==0.1.2
configdict==2.5.2
contourpy==1.0.7
cryptography @ file:///croot/cryptography_1673298753778/work
cv_bridge==1.13.1
cycler==0.11.0
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
descartes==1.1.0
diagnostic_updater==1.9.7
docker-pycreds==0.4.0
exceptiongroup==1.1.0
executing==1.2.0
fastjsonschema==2.16.2
fire==0.5.0
flake8==6.0.0
flit_core @ file:///opt/conda/conda-bld/flit-core_1644941570762/work/source/flit_core
fonttools==4.38.0
fqdn==1.5.1
fuzzywuzzy==0.18.0
gencpp==0.6.6
geneus==2.2.6
genlisp==0.4.16
genmsg==0.5.17
gennodejs==2.0.1
genpy==0.6.16
gitdb==4.0.10
GitPython==3.1.30
google-auth==2.16.0
google-auth-oauthlib==0.4.6
graphviz==0.20.1
grpcio==1.51.1
idna @ file:///croot/idna_1666125576474/work
image_geometry==1.13.1
imageio==2.25.0
importlib-metadata==6.0.0
importlib-resources==5.10.2
iniconfig==2.0.0
ipykernel==6.21.1
ipython==8.10.0
ipython-genutils==0.2.0
ipywidgets==8.0.4
isoduration==20.11.0
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
joint_state_publisher==1.12.15
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.5.0
jupyter-events==0.6.3
jupyter_client==8.0.2
jupyter_core==5.2.0
jupyter_server==2.2.1
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.5
kdl-parser-py==1.13.3
kiwisolver==1.4.4
llvmlite==0.36.0
lyft-dataset-sdk==0.0.8
Markdown==3.4.1
markdown-it-py==2.1.0
MarkupSafe==2.1.2
matplotlib==3.6.3
matplotlib-inline==0.1.6
mccabe==0.7.0
mdurl==0.1.2
message_filters==1.14.13
mistune==2.0.5
mkl-fft==1.3.1
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626186064646/work
mkl-service==2.4.0
mmcls==0.25.0
mmcv-full==1.6.2
mmdet==2.28.1
-e git+https://github.com/adastec/mmdetection3d.git@c3f8502468e2f780b2ce45d4efdbc0a2ab00cea9#egg=mmdet3d
mmsegmentation==0.30.0
model-index==0.1.11
mypy-extensions==1.0.0
nbclassic==0.5.1
nbclient==0.7.2
nbconvert==7.2.9
nbformat==5.7.3
nest-asyncio==1.5.6
networkx==2.2
notebook==6.5.2
notebook_shim==0.2.2
numba==0.53.0
numpy==1.19.5
nuscenes-devkit==1.1.9
oauthlib==3.2.2
opencv-python==4.7.0.68
openmim==0.3.6
ordered-set==4.1.0
packaging==23.0
pandas==1.4.4
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.11.0
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
pkgutil_resolve_name==1.3.10
platformdirs==3.0.0
plotly==5.13.0
pluggy==1.0.0
plyfile==0.7.4
prettytable==3.6.0
prometheus-client==0.16.0
promise==2.3
prompt-toolkit==3.0.36
protobuf==3.20.3
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.6
pycodestyle==2.10.0
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pyflakes==3.0.1
Pygments==2.14.0
pyOpenSSL @ file:///opt/conda/conda-bld/pyopenssl_1643788558760/work
pyparsing==3.0.9
pyquaternion==0.9.9
pyrsistent==0.19.3
PySocks @ file:///tmp/build/80754af9/pysocks_1605305779399/work
pytest==7.2.1
python-dateutil==2.8.2
python-json-logger==2.0.4
python_qt_binding==0.4.4
pytz==2022.7.1
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==25.0.0
qt-dotgraph==0.4.2
qt-gui==0.4.2
qt-gui-py-common==0.4.2
qtconsole==5.4.0
QtPy==2.3.0
requests @ file:///opt/conda/conda-bld/requests_1657734628632/work
requests-oauthlib==1.3.1
resource_retriever==1.12.7
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.3.1
rosboost-cfg==1.14.9
rosclean==1.14.9
roscreate==1.14.9
rosgraph==1.14.13
roslaunch==1.14.13
roslib==1.14.9
roslint==0.11.2
roslz4==1.14.13
rosmake==1.14.9
rosmaster==1.14.13
rosparam==1.14.13
rospy==1.14.13
rostest==1.14.13
rosunit==1.14.9
rqt_gui==0.5.3
rqt_gui_py==0.5.3
rqt_py_console==0.4.8
rqt_runtime_monitor==0.5.7
rqt_shell==0.4.9
rqt_top==0.4.8
rqt_web==0.4.8
rsa==4.9
scikit-image==0.19.3
scikit-learn==1.2.1
scipy==1.10.0
Send2Trash==1.8.0
sensor-msgs==1.12.8
sentry-sdk==1.15.0
shapely==2.0.1
shortuuid==1.0.11
six @ file:///tmp/build/80754af9/six_1644875935023/work
smach==2.0.1
smclib==1.8.5
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.3.2.post1
stack-data==0.6.2
tabulate==0.9.0
tenacity==8.2.1
tensorboard==2.12.0
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
termcolor==2.2.0
terminado==0.17.1
terminaltables==3.1.10
tf2_py==0.6.5
threadpoolctl==3.1.0
tifffile==2023.2.3
tinycss2==1.2.1
tomli==2.0.1
topic_tools==1.14.13
torch==1.12.1
torchaudio==0.12.1
torchvision==0.13.1
torchviz==0.0.2
tornado==6.2
tqdm==4.64.1
traitlets==5.9.0
trimesh==2.35.39
typing_extensions @ file:///croot/typing_extensions_1669924550328/work
urdfdom-py==0.4.6
uri-template==1.2.0
urllib3 @ file:///croot/urllib3_1673575502006/work
wandb==0.12.10
watchdog==2.2.1
wcwidth==0.2.6
webcolors==1.12
webencodings==0.5.1
websocket-client==1.5.1
Werkzeug==2.2.2
widgetsnbextension==4.0.5
xacro==1.13.18
yapf==0.32.0
yaspin==2.3.0
zipp==3.13.0
